requirements.txt
Flask==3.0.0
python-dotenv>=1.0.0
claude-code-sdk
anyio>=4.0.0
browser-use==0.5.9
typing-extensions>=4.0.0
